pyproject.toml
[project]
name = "notes-recall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-text-splitters>=1.1.2",
    "openai>=1.0.0",
    "pydantic>=2.13.3",
    "pylint>=4.0.5",
    "python-dotenv>=1.2.2",
]
